+++ tests/test_diff.py
@@ -1,6 +1,6 @@
 import itertools
 
-from openelm.util.diff_eval import apply_diff, parse_line_info, replace_text, split_diff, verify_diff, DiffState
+from openelm.utils.diff_eval import apply_diff, parse_line_info, replace_text, split_diff, verify_diff, DiffState
 
 
 def test_diff():
